+++ rollingops/CHANGELOG.md
@@ -1,3 +1,11 @@
+# 1.1.0 - 20 May 2026
+
+Extend the `RollingOpsManager` is_waiting_* helpers to receive a unit name.
+
+# 1.0.1 - 04 May 2026
+
+Fix the `ModelError` messages generated during rollback operations.
+
 # 1.0.0 - 28 April 2026
 
 Initial release of `charmlibs-rollingops` library